extracts security problems client
