+++ internal/dynatrace/api.go
@@ -51,60 +51,3 @@ type MetricsQueryResult struct {
 	NextPageKey string                    `json:"nextPageKey"`
 	Result      []MetricQueryResultValues `json:"result"`
 }
-
-// SecurityProblem Problem Detail returned by /api/v2/securityProblems
-type SecurityProblem struct {
-	SecurityProblemID    string `json:"securityProblemId"`
-	DisplayID            int    `json:"displayId"`
-	State                string `json:"state"`
-	VulnerabilityID      string `json:"vulnerabilityId"`
-	VulnerabilityType    string `json:"vulnerabilityType"`
-	FirstSeenTimestamp   int    `json:"firstSeenTimestamp"`
-	LastUpdatedTimestamp int    `json:"lastUpdatedTimestamp"`
-	RiskAssessment       struct {
-		RiskCategory string `json:"riskCategory"`
-		RiskScore    struct {
-			Value int `json:"value"`
-		} `json:"riskScore"`
-		Exposed                bool `json:"exposed"`
-		SensitiveDataAffected  bool `json:"sensitiveDataAffected"`
-		PublicExploitAvailable bool `json:"publicExploitAvailable"`
-	} `json:"riskAssessment"`
-	ManagementZones      []string `json:"managementZones"`
-	VulnerableComponents []struct {
-		ID                          string   `json:"id"`
-		DisplayName                 string   `json:"displayName"`
-		FileName                    string   `json:"fileName"`
-		NumberOfVulnerableProcesses int      `json:"numberOfVulnerableProcesses"`
-		VulnerableProcesses         []string `json:"vulnerableProcesses"`
-	} `json:"vulnerableComponents"`
-	VulnerableEntities  []string `json:"vulnerableEntities"`
-	ExposedEntities     []string `json:"exposedEntities"`
-	SensitiveDataAssets []string `json:"sensitiveDataAssets"`
-	AffectedEntities    struct {
-		Applications []struct {
-			ID                          string   `json:"id"`
-			NumberOfVulnerableProcesses int      `json:"numberOfVulnerableProcesses"`
-			VulnerableProcesses         []string `json:"vulnerableProcesses"`
-		} `json:"applications"`
-		Services []struct {
-			ID                          string   `json:"id"`
-			NumberOfVulnerableProcesses int      `json:"numberOfVulnerableProcesses"`
-			VulnerableProcesses         []string `json:"vulnerableProcesses"`
-		} `json:"services"`
-		Hosts []struct {
-			ID                          string   `json:"id"`
-			NumberOfVulnerableProcesses int      `json:"numberOfVulnerableProcesses"`
-			VulnerableProcesses         []string `json:"vulnerableProcesses"`
-		} `json:"hosts"`
-		Databases []string `json:"databases"`
-	} `json:"affectedEntities"`
-}
-
-// SecurityProblemQueryResult Result of/api/v2/securityProblems
-type SecurityProblemQueryResult struct {
-	TotalCount       int               `json:"totalCount"`
-	PageSize         int               `json:"pageSize"`
-	NextPageKey      string            `json:"nextPageKey"`
-	SecurityProblems []SecurityProblem `json:"securityProblems"`
-}
