+++ internal/sli/dynatrace.go
@@ -111,28 +111,6 @@ func (ph *Handler) loadDynatraceDashboard(keptnEvent adapter.EventContentAdapter
 	return dynatraceDashboard, dashboard, nil
 }
 
-// executeGetDynatraceSecurityProblems Calls the /securityProblems/ API call to retrieve the list of security problems for that timeframe.
-// It returns a SecurityProblemQueryResult object on success, an error otherwise.
-func (ph *Handler) executeGetDynatraceSecurityProblems(problemQuery string, startUnix time.Time, endUnix time.Time) (*dynatrace.SecurityProblemQueryResult, error) {
-	body, err := ph.dtClient.Get(
-		fmt.Sprintf("/api/v2/securityProblems?from=%s&to=%s&%s",
-			common.TimestampToString(startUnix),
-			common.TimestampToString(endUnix),
-			problemQuery))
-	if err != nil {
-		return nil, err
-	}
-
-	// parse response json
-	var result dynatrace.SecurityProblemQueryResult
-	err = json.Unmarshal(body, &result)
-	if err != nil {
-		return nil, err
-	}
-
-	return &result, nil
-}
-
 // executeMetricAPIDescribe Calls the /metrics/<metricID> API call to retrieve Metric Definition Details.
 func (ph *Handler) executeMetricAPIDescribe(metricId string) (*dynatrace.MetricDefinition, error) {
 	body, err := ph.dtClient.Get("/api/v2/metrics/" + metricId)
@@ -456,7 +434,7 @@ func (ph *Handler) processOpenSecurityProblemTile(securityProblemSelector string
 	}
 
 	// Step 1: Query the Dynatrace API to get the number of actual problems matching that query and timeframe
-	problemQueryResult, err := ph.executeGetDynatraceSecurityProblems(problemQuery, startUnix, endUnix)
+	problemQueryResult, err := dynatrace.NewSecurityProblemsClient(ph.dtClient).GetByQuery(problemQuery, startUnix, endUnix)
 	if err != nil {
 		return nil, "", "", nil, err
 	}
@@ -1290,9 +1268,9 @@ func (ph *Handler) executeSecurityProblemQuery(metricsQuery string, startUnix ti
 	}
 
 	problemQuery := querySplits[1]
-	problemQueryResult, err := ph.executeGetDynatraceSecurityProblems(problemQuery, startUnix, endUnix)
+	problemQueryResult, err := dynatrace.NewSecurityProblemsClient(ph.dtClient).GetByQuery(problemQuery, startUnix, endUnix)
 	if err != nil {
-		return 0, fmt.Errorf("Error executing Dynatrace Security Problem v2 Query %v", err)
+		return 0, err
 	}
 
 	return float64(problemQueryResult.TotalCount), nil
